Reorganize paths to separate auth paths (e.g. login/logout) from regular user action paths (e.g. saving/deleting profile) - Update server.js app.use() statements: /auth for auth actions, and /identity for regular user actions - Separated auth routes and regular user action routes - Updated imports & exports
