+++ backend/src/apps/users/entry-points/api/index.js
@@ -1,24 +1,4 @@
-import express from "express";
-import {
-  getGoogleAuth,
-  getGithubAuth,
-  getGoogleAuthCallback,
-  getGithubAuthCallback,
-  getAuthSuccess,
-  postLogout,
-  patchUser,
-  deleteUser,
-} from "./users-controller.js";
-import makeExpressCallback from "../../express-callback/index.js";
-import authMiddleware from "../../auth-middleware/index.js";
+import authRoutes from "./auth-routes.js";
+import userRoutes from "./user-routes.js";
 
-const router = express.Router();
-router.get("/google", getGoogleAuth);
-router.get("/github", getGithubAuth);
-router.get(process.env.GOOGLE_CALLBACK_URL, getGoogleAuthCallback);
-router.get(process.env.GITHUB_CALLBACK_URL, getGithubAuthCallback);
-router.get("/success", makeExpressCallback(getAuthSuccess));
-router.post("/logout", makeExpressCallback(postLogout));
-router.patch("/user", authMiddleware(true), makeExpressCallback(patchUser));
-router.delete("/user", authMiddleware(), makeExpressCallback(deleteUser));
-export default router;
+export { authRoutes, userRoutes };
